Image: Move scaling functions to seperate header
